requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic

livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-google

groq
openai

websockets
aiohttp

numpy